+++ eslintConfig.js
@@ -58,11 +58,5 @@ module.exports = {
     'operator-linebreak': 'off',
     indent: 'off',
     curly: 'off',
-    '@typescript-eslint/explicit-function-return-type': 'off',
-    '@typescript-eslint/explicit-module-boundary-types': 'off',
-    '@typescript-eslint/no-explicit-any': ['error', { ignoreRestArgs: true }],
-    '@typescript-eslint/no-var-requires': 'off',
-    '@typescript-eslint/no-use-before-define': 'off',
-    '@typescript-eslint/no-empty-function': 'off',
   },
 };