+++ packages/plugin-image/web/src/typeMapper.js
@@ -1,5 +1,4 @@
 import ImageViewer from './image-viewer';
-import { IMAGE_TYPE_LEGACY, IMAGE_TYPE } from './types';
 import { sizeClassName, alignmentClassName } from './classNameStrategies';
 
 const imageRenderDescriptor = {
@@ -11,6 +10,6 @@ const imageRenderDescriptor = {
 };
 
 export const typeMapper = () => ({
-  [IMAGE_TYPE_LEGACY]: imageRenderDescriptor,
-  [IMAGE_TYPE]: imageRenderDescriptor,
+  IMAGE: imageRenderDescriptor,
+  'wix-draft-plugin-image': imageRenderDescriptor,
 });