+++ packages/ricos-editor/web/src/modals/RicosModal.tsx
@@ -1,4 +1,3 @@
-/* eslint-disable react/prop-types */
 import React, { Fragment, ComponentType, Children, FunctionComponent } from 'react';
 import EditorModalProvider from './EditorModalProvider';