+++ rollup.plugins.js
@@ -159,7 +159,16 @@ const visualizer = () => {
   });
 };
 
-let _plugins = [svgr(), resolveAlias(), resolve(), copy(), babel(), commonjs(), json()];
+let _plugins = [
+  svgr(),
+  resolveAlias(),
+  resolve(),
+  copy(),
+  babel(),
+  commonjs(),
+  json(),
+  typescript(),
+];
 
 if (!IS_DEV_ENV) {
   _plugins = [..._plugins, replace(), uglify()];
@@ -167,10 +176,6 @@ if (!IS_DEV_ENV) {
 
 if (process.env.MODULE_ANALYZE_EDITOR || process.env.MODULE_ANALYZE_VIEWER) {
   _plugins = [..._plugins, visualizer()];
-}
-
-if (process.env.TS) {
-  _plugins = [..._plugins, typescript()];
 }
 
 const plugins = shouldExtractCss => {